openMask/requirements.txt
pillow
openCV-python
numpy